requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.12
ciscoconfparse==1.6.40
colorama==0.4.4
cryptography==3.4.4
dnspython==2.2.1
future==0.18.2
idna==3.3
Jinja2==3.1.1
junos-eznc==2.6.3
loguru==0.6.0
lxml==4.8.0
MarkupSafe==2.1.1
napalm==3.4.0
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.3.3
ntc-templates==1.6.0
packaging==21.3
paramiko==2.7.2
passlib==1.7.4
pycparser==2.20
pyeapi==0.8.4
PyNaCl==1.4.0
pynetbox==6.6.1
pyparsing==3.0.7
pyserial==3.5
python-certifi-win32==1.6
PyYAML==5.4.1
requests==2.27.1
scp==0.13.3
setuptools-scm==6.4.2
six==1.15.0
tabulate==0.8.9
tenacity==6.3.1
textfsm==1.1.0
toml==0.10.2
tomli==2.0.1
transitions==0.8.11
urllib3==1.26.9
win32-setctime==1.1.0
wrapt==1.14.0
yamlordereddictloader==0.4.0
